requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
scikit-image==0.21.0
scipy==1.11.3
gdown==4.7.1
requests==2.31.0
tqdm==4.66.1